postgres CRUD with http messages

+++ postgres/postgres_runner.py
@@ -44,12 +44,16 @@ class PostgresRunner:
         all_tables = self.list_all_tables()
         return any(input_table_name.lower() in table for table in all_tables)
 
+    def existing_query(self, query_dict: dict):
+        all_queries = self.list_all_queries()
+        return any(query_dict['query_name'].lower() in query for query in all_queries)
+
     def create_flight_query_table(self):
         if self.existing_table(self.table_name):
             print(f"Could not create {self.table_name} because that table already exists")
             return
         sql = 'create table FlightQuery (id serial primary key, fly_from varchar(20), fly_to varchar(20), ' \
-              'date_from varchar(20), date_to varchar(20), query_limit INT)'
+              'date_from varchar(20), date_to varchar(20), query_limit INT, query_name varchar(20))'
         self.run_sql_query(sql)
 
     def delete_flight_query_table(self):
@@ -61,56 +65,59 @@ class PostgresRunner:
         self.run_sql_query(sql)
         return self.show_query_results()
 
-    def insert_flight_query(self, query_dict: dict):
-        sql = f"insert into {self.table_name} values ({self.query_amount}, '{query_dict['fly_from']}'," \
-              f" '{query_dict['fly_to']}', " \
-              f"'{query_dict['date_from']}', '{query_dict['date_to']}', {query_dict['query_limit']}) "
-        self.run_sql_query(sql)
-
     def get_flight_query_id(self, query_dict: dict):
-        sql = f"select id from {self.table_name} where fly_from = '{query_dict['fly_from']}'" \
-              f" and fly_to = '{query_dict['fly_to']}' and " \
-              f"date_from = '{query_dict['date_from']}' and date_to = '{query_dict['date_to']}' and " \
-              f"query_limit = {query_dict['query_limit']}"
+        sql = f"select id from {self.table_name} where query_name = {query_dict['query_name']}"
         self.run_sql_query(sql)
         return self.show_query_results()[0][0]
 
-    def modify_flight_query(self, query_dict: dict):
+    def flight_query_create(self, query_dict: dict) -> bool:
+        if self.existing_query(query_dict):
+            return False
+        sql = f"insert into {self.table_name} values ({self.query_amount}, '{query_dict['fly_from']}'," \
+              f" '{query_dict['fly_to']}', " \
+              f"'{query_dict['date_from']}', '{query_dict['date_to']}', {query_dict['query_limit']}," \
+              f"{query_dict['query_name']}) "
+        self.run_sql_query(sql)
+        return True
+
+    def flight_query_read(self, query_dict: dict) -> bool or list:
+        if not self.existing_query(query_dict):
+            return False
+        query_id = self.get_flight_query_id(query_dict)
+        sql = f"select * from {self.table_name} where id={query_id}"
+        self.run_sql_query(sql)
+        return self.show_query_results()
+
+    def flight_query_update(self, query_dict: dict) -> bool:
+        if not self.existing_query(query_dict):
+            return False
         query_id = self.get_flight_query_id(query_dict)
         sql = f"update {self.table_name} " \
               f"set fly_from='{query_dict['fly_from']}', fly_to='{query_dict['fly_to']}', " \
               f"date_from='{query_dict['date_from']}', date_to='{query_dict['date_to']}'," \
               f"query_limit={query_dict['query_limit']} WHERE ID={query_id}"
         self.run_sql_query(sql)
+        return True
 
-    def delete_flight_query(self, query_dict: dict):
+    def flight_query_delete(self, query_dict: dict) -> bool:
+        if not self.existing_query(query_dict):
+            return False
         query_id = self.get_flight_query_id(query_dict)
         sql = f"delete from {self.table_name} where id={query_id}"
         self.run_sql_query(sql)
+        return True
 
-
-# sql = 'create table cidade (id serial primary key, nome varchar(100), uf varchar(2))'
-# cur.execute(sql)
-# sql = "insert into cidade values (default,'São Paulo,'SP')"
-# cur.execute(sql)
-# con.commit()
-# cur.execute('select * from cidade')
-# recset = cur.fetchall()
-# for rec in recset:
-#     print (rec)
 
 def __main():
-    runner = PostgresRunner()
-    # runner.delete_flight_query_table()
-    # runner.create_flight_query_table()
+    pgr = PostgresRunner()
     aux = {"fly_from": "FOR", "fly_to": "SP", "date_from": "01/07/2022", "date_to": "01/07/2023", "query_limit": 50}
-    runner.delete_flight_query(aux)
-    print(runner.list_all_queries())
+    pgr.delete_flight_query_table()
+    print(pgr.list_all_queries())
     # print(runner.list_all_schemas())
     # print(runner.existing_table("FlightQuery"))
     # print(runner.existing_table())
     # runner.create_flight_query_table()
-    runner.close_con()
+    pgr.close_con()
 
 
 if __name__ == "__main__":
